refactor: use new auth provider
